Added `SyncTemplateRepository` - partially synchronous implementation of the `TemplateRepositoryLike`
Summary:
This is partially synchronous version of the `AsyncTemplateRepository` class. For more details on how this
repository is saving the data in the Zookeeper see the comments to the `AsyncTemplateRepository` class.

However, unlike it's asynchronous counterpart, this class reads the contents of the store on it's initialization
keeping everything in memory using a `PathTrie`. This allows for synchronous reading of the templates via
additional methods `readSync`, `contentsSync` and `existsSync`. In fact, the asynchronous versions of these
methods are calling their synchronous counterparts wrapping the calls into a future.

All the reading (synchronous) methods are served directly from the memory, all the writing calls e.g. `create`,
`delete` are changing the underlying storage first and the in-memory copy afterwards.